new: add benchmark
initial test using pyperf + timeit... see how long this takes :o

test: something wrong with yaml?

test: okay.. revert more :(

test: okay.. so benchmark was the issue...

is it the python commands?

fix: timeit is native package i guess.

fix: was it the new line in the middle of the string?

fix: with + for can't be on the same line :(

fix: \ for multiline string?

test fix: > instead of |